Tests for the Google login
